Handle Registration/Deletion of Producers and Consumers
